Refactor #18 : TimeScreen.kt 파일 지우고 TimeRoute.kt안으로 내용 옮김
